Create custom filters

+++ section_10/spring-sec-section-10/src/main/java/com/ttdat/eazybank/repository/AccountTransactionsRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 @Repository
 public interface AccountTransactionsRepository extends CrudRepository<AccountTransactions, Long> {
-	List<AccountTransactions> findByCustomerIdOrderByTransactionDtDesc(int customerId);
+	List<AccountTransactions> findByCustomerIdOrderByTransactionDtDesc(long customerId);
 }
