+++ section_10/spring-sec-section-10/src/main/java/com/ttdat/eazybank/repository/CardsRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface CardsRepository extends CrudRepository<Cards, Long> {
-	List<Cards> findByCustomerId(int customerId);
+	List<Cards> findByCustomerId(long customerId);
 }
