+++ section_10/spring-sec-section-10/src/main/java/com/ttdat/eazybank/repository/CustomerRepository.java
@@ -4,9 +4,10 @@ import com.ttdat.eazybank.model.Customer;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface CustomerRepository extends CrudRepository<Customer, Long> {
+public interface CustomerRepository extends CrudRepository<Customer, Integer> {
     Optional<Customer> findByEmail(String email);
 }
